chainntnfs: modify GetTestTxidAndScript to generate new P2WPKH scripts
In this commit, we modify GetTestTxidAndScript to generate new P2WPKH
scripts. This is needed for the following reasons:

1. Properly re-derive the script of an output from the input spending
it when detecting script spends. The previous P2PKH scripts are not
supported by this.

2. Properly test confirmations and spends of scripts on-chain through
the set of interface-level test cases.

The test chains also need to be extended in order to ensure that segwit
has been activated.