routing: remove k-shortest route cache
This commit removes route caching for QueryRoutes. The cache is broken
because it doesn't take into account the fee limit and final cltv delta
of a QueryRoutes request.

The original reason for adding the cache is described in the commit log
as to speed up payments. However, in the SendPayment paths the cache
isn't used. It is therefore unlikely that the cache adds much real use
benefit.