channeldb+invoices: add invoice htlcs
This commit adds a set of htlcs to the Invoice struct and
serializes/deserializes this set to/from disk. It is a preparation for
accurate invoice accounting across restarts of lnd.

A migration is added for the invoice htlcs.

In addition to these changes, a separate final cltv delta invoice field
is created and populated. Previously it was required to decode this from
the stored payment request. The reason to create a combined commit is to
prevent writing two separate migrations.