channeldb+routing: add FetchPayment method
This method is used to fetch a payment and all HTLC attempt that have
been made for that payment. It will both be used to resume inflight
attempts, and to fetch the final outcome of previous attempts.

We also prepare the mock control tower for MPP, by letting it have
multiple attempts in flight for a given payment hash, settling and
failing them individually.