channeldb+invoices: process legacy keysend as AMP payment
This commit automatically generates a random payment address and set id
for any new keysend payments. With the addition of the payment address
index, we now enforce uniqueness of payment addresses across all
invoices. As a result, if you run master currently you can only
accept one keysend payment since they all share a zero-value payment
address. If we detect that this is a replayed HTLC, we will extract the
payment address and set id generated on the prior pass.